+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## UNRELEASED
 
+## [1.7.0] - 2023-07-12
 ### Added
 - Make CAS extra attributes available in `Cas::Authentication::User`